+++ status_check_stage_bridge/constants.rb
@@ -1,7 +1,7 @@
 module StatusCheckGG
   module StageBridge
-    VERSION = '0.1.0-beta.2'
-    BRIDGE_SCHEMA_VERSION = 2
+    VERSION = '0.1.0-beta.3'
+    BRIDGE_SCHEMA_VERSION = 3
 
     MODEL_DICTIONARY = 'status_check_stage_bridge_model'
     ENTITY_DICTIONARY = 'status_check_stage_bridge_entity'
@@ -11,6 +11,6 @@ module StatusCheckGG
     ROLE_STAGE_REFERENCE = 'stage_reference'
     ROLE_PROP = 'prop'
 
-    TRANSFORM_PROFILE = 'practisim-0.01005m-source-flip-y-negate-yaw-ground-offset-v2'
+    TRANSFORM_PROFILE = 'practisim-0.01005m-source-flip-y-negate-yaw-ground-offset-adjustable-endpoint-v3'
   end
 end